+++ psutil/arch/netbsd/mem.c
@@ -32,6 +32,7 @@ psutil_virtual_mem(PyObject *self, PyObject *args) {
     struct uvmexp_sysctl uv;
     int mib[] = {CTL_VM, VM_UVMEXP2};
     long pagesize = psutil_getpagesize();
+    unsigned long long available;
 
     size = sizeof(uv);
     if (sysctl(mib, 2, &uv, &size, NULL, 0) < 0) {
@@ -39,13 +40,17 @@ psutil_virtual_mem(PyObject *self, PyObject *args) {
         return NULL;
     }
 
-    return Py_BuildValue("KKKKKK",
+    // follow zabbix
+    available = uv.inactive + uv.execpages + uv.filepages + uv.free;
+    return Py_BuildValue(
+        "KKKKKKK",
         (unsigned long long) uv.npages << uv.pageshift,  // total
         (unsigned long long) uv.free << uv.pageshift,  // free
         (unsigned long long) uv.active << uv.pageshift,  // active
         (unsigned long long) uv.inactive << uv.pageshift,  // inactive
         (unsigned long long) uv.wired << uv.pageshift,  // wired
-        (unsigned long long) (uv.filepages + uv.execpages) * pagesize  // cached
+        (unsigned long long) (uv.filepages + uv.execpages) * pagesize,  // cached
+        available << uv.pageshift  // available
     );
 }